Add flux-pattern. User login state done with flux-pattern
LoginStore keeps tabs on logged in user. Login is emitted via event and dispatcher to react components. LoginStore doesnt check if a login session exists yet, so refreshing will remove the login state. The link to edit a blogpost is only visible for logged in users.

Added couple of utility modules to package.json that were used by the todoapp facebook example of flux.